requirements.txt
tensorflow==1.10
numpy==1.14.5
argparse==1.1
Pillow==5.2.0